+++ packages/build-info/src/build-systems/rush.ts
@@ -1,0 +1,7 @@
+import { BaseBuildTool } from './build-system.js'
+
+export class Rush extends BaseBuildTool {
+  id = 'rush'
+  name = 'Rush'
+  configFiles = ['rush.json']
+}