+++ packages/build-info/src/index.ts
@@ -1,1 +1,2 @@
-export { getBuildInfo } from './get-build-info.js'
+export * from './file-system.js'
+export * from './project.js'